+++ src/main/php/lang/ast/nodes/CastExpression.class.php
@@ -13,5 +13,5 @@ class CastExpression extends Node {
   }
 
   /** @return iterable */
-  public function children() { return [$this->expression]; }
+  public function children() { return [&$this->expression]; }
 }