+++ src/main/php/lang/ast/nodes/ContinueStatement.class.php
@@ -12,7 +12,5 @@ class ContinueStatement extends Node {
   }
 
   /** @return iterable */
-  public function children() {
-    return $this->expression ? [$this->expression] : [];
-  }
+  public function children() { return $this->expression ? [&$this->expression] : []; }
 }