+++ src/test/php/lang/ast/unittest/parse/MembersTest.class.php
@@ -3,20 +3,23 @@
 use lang\ast\Type;
 use lang\ast\nodes\{
   Annotations,
+  Block,
   ClassDeclaration,
   Constant,
   Expression,
+  Hook,
   InstanceExpression,
   InvokeExpression,
   Literal,
   Method,
   Property,
+  ReturnStatement,
   ScopeExpression,
   Signature,
   Variable,
   Parameter
 };
-use lang\ast\types\{IsFunction, IsLiteral, IsNullable, IsUnion, IsValue};
+use lang\ast\types\{IsFunction, IsLiteral, IsNullable, IsUnion, IsValue, IsGeneric};
 use test\{Assert, Test, Values};
 
 class MembersTest extends ParseTest {
@@ -137,6 +140,71 @@ class MembersTest extends ParseTest {
   }
 
   #[Test]
+  public function property_with_get_and_set_hooks() {
+    $class= new ClassDeclaration([], new IsValue('\\A'), null, [], [], null, null, self::LINE);
+    $prop= new Property(['public'], 'a', null, null, null, null, self::LINE);
+    $return= new ReturnStatement(new Literal('"Hello"', self::LINE), self::LINE);
+    $parameter= new Parameter('value', null, null, false, false, null, null, null, self::LINE);
+    $prop->hooks['get']= new Hook([], 'get', new Block([$return], self::LINE), false, null, self::LINE);
+    $prop->hooks['set']= new Hook([], 'set', new Block([], self::LINE), false, $parameter, self::LINE);
+    $class->declare($prop);
+
+    $this->assertParsed([$class], 'class A { public $a { get { return "Hello"; } set($value) { } } }');
+  }
+
+  #[Test]
+  public function property_with_short_get_hook() {
+    $class= new ClassDeclaration([], new IsValue('\\A'), null, [], [], null, null, self::LINE);
+    $prop= new Property(['public'], 'a', null, null, null, null, self::LINE);
+    $prop->hooks['get']= new Hook([], 'get', new Literal('"Hello"', self::LINE), false, null, self::LINE);
+    $class->declare($prop);
+
+    $this->assertParsed([$class], 'class A { public $a { get => "Hello"; } }');
+  }
+
+  #[Test]
+  public function property_with_abbreviated_get_hook() {
+    $class= new ClassDeclaration([], new IsValue('\\A'), null, [], [], null, null, self::LINE);
+    $prop= new Property(['public'], 'a', null, null, null, null, self::LINE);
+    $prop->hooks['get']= new Hook([], 'get', new Literal('"Hello"', self::LINE), false, null, self::LINE);
+    $class->declare($prop);
+
+    $this->assertParsed([$class], 'class A { public $a => "Hello"; }');
+  }
+
+  #[Test]
+  public function property_with_typed_hook() {
+    $class= new ClassDeclaration([], new IsValue('\\A'), null, [], [], null, null, self::LINE);
+    $prop= new Property(['public'], 'a', null, null, null, null, self::LINE);
+    $parameter= new Parameter('value', new IsLiteral('string'), null, false, false, null, null, null, self::LINE);
+    $prop->hooks['set']= new Hook([], 'set', new Block([], self::LINE), false, $parameter, self::LINE);
+    $class->declare($prop);
+
+    $this->assertParsed([$class], 'class A { public $a { set(string $value) { } } }');
+  }
+
+  #[Test]
+  public function property_with_final_hook() {
+    $class= new ClassDeclaration([], new IsValue('\\A'), null, [], [], null, null, self::LINE);
+    $prop= new Property(['public'], 'a', null, null, null, null, self::LINE);
+    $parameter= new Parameter('value', new IsLiteral('string'), null, false, false, null, null, null, self::LINE);
+    $prop->hooks['set']= new Hook(['final'], 'set', new Block([], self::LINE), false, $parameter, self::LINE);
+    $class->declare($prop);
+
+    $this->assertParsed([$class], 'class A { public $a { final set(string $value) { } } }');
+  }
+
+  #[Test]
+  public function abstract_property_with_hook() {
+    $class= new ClassDeclaration(['abstract'], new IsValue('\\A'), null, [], [], null, null, self::LINE);
+    $prop= new Property(['public', 'abstract'], 'a', null, null, null, null, self::LINE);
+    $prop->hooks['set']= new Hook([], 'set', null, false, null, self::LINE);
+    $class->declare($prop);
+
+    $this->assertParsed([$class], 'abstract class A { public abstract $a { set; } }');
+  }
+
+  #[Test]
   public function instance_property_access() {
     $this->assertParsed(
       [new InstanceExpression(new Variable('a', self::LINE), new Literal('member', self::LINE), self::LINE)],
@@ -230,6 +298,14 @@ class MembersTest extends ParseTest {
   }
 
   #[Test]
+  public function generic_class_resolution() {
+    $this->assertParsed(
+      [new ScopeExpression(new IsGeneric(new IsValue('\\A'), [new IsValue('\\T')]), new Literal('class', self::LINE), self::LINE)],
+      'A<T>::class;'
+    );
+  }
+
+  #[Test]
   public function instance_method_invocation() {
     $this->assertParsed(
       [new InvokeExpression(
@@ -303,5 +379,22 @@ class MembersTest extends ParseTest {
     $class->declare(new Constant([], 'I', new Type('string'), new Literal('"i"', self::LINE), null, null, self::LINE));
 
     $this->assertParsed([$class], 'class A { const int T = 1, S = 2, string I = "i"; }');
+  }
+
+  #[Test]
+  public function asymmetric_property() {
+    $class= new ClassDeclaration([], new IsValue('\\A'), null, [], [], null, null, self::LINE);
+    $class->declare(new Property(['public', 'private(set)'], 'a', new Type('int'), null, null, null, self::LINE));
+
+    $this->assertParsed([$class], 'class A { public private(set) int $a; }');
+  }
+
+  #[Test]
+  public function asymmetric_property_as_constructor_argument() {
+    $params= [new Parameter('a', new IsLiteral('int'), null, false, false, ['private(set)'], null, null, self::LINE)];
+    $class= new ClassDeclaration([], new IsValue('\\A'), null, [], [], null, null, self::LINE);
+    $class->declare(new Method(['public'], '__construct', new Signature($params, null, false, self::LINE), [], null, null, self::LINE));
+
+    $this->assertParsed([$class], 'class A { public function __construct(private(set) int $a) { } }');
   }
 }